requirements.txt
altgraph==0.17
arabic-reshaper==2.1.3
asgiref==3.3.4
auto-py-to-exe==2.9.0
autopep8==1.5.7
Babel==2.9.1
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-money==2.0
django-seed==0.2.2
django-templated-mail==1.1.1
django-unused-media==0.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-writable-nested==0.6.3
drf-yasg==1.20.0
Eel==0.12.4
Faker==8.8.1
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==20.9
Pillow==8.2.0
psycopg2-binary==2.9.1
py-moneyed==1.2
pycodestyle==2.7.0
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
webencodings==0.5.1
whichcraft==0.6.1
xhtml2pdf==0.2.5
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
